# pkg/swagger/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This API provides programmatic access to sending SMS messages on Android devices. Features include sending SMS, checking message status, device management, webhook configuration, and system health checks.",
        "title": "SMS Gateway for Android™ API",
        "contact": {
            "name": "SMSGate Support",
            "email": "support@sms-gate.app"
        },
        "version": "{APP_VERSION}"
    },
    "host": "api.sms-gate.app",
    "paths": {
...
        "/mobile/v1/device": {
            "get": {
                "description": "Returns device information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "Get device information",
                "responses": {
                    "200": {
                        "description": "Device information",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.MobileDeviceResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiAuth": []
                    },
                    {
                        "UserCode": []
                    },
                    {
                        "ServerKey": []
                    }
                ],
                "description": "Registers new device for new or existing user. Returns user credentials only for new users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "Register device",
                "parameters": [
                    {
                        "description": "Device registration request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/smsgateway.MobileRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Device registered",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.MobileRegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (private mode only)",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "MobileToken": []
                    }
                ],
                "description": "Updates push token for device",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "Update device",
                "parameters": [
                    {
                        "description": "Device update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/smsgateway.MobileUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully updated"
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (wrong device ID)",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/mobile/v1/events": {
            "get": {
                "security": [
                    {
                        "MobileToken": []
                    }
                ],
                "description": "Returns events stream for a device",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "Device",
                    "Events"
                ],
                "summary": "Get events",
                "responses": {
                    "200": {
                        "description": "Event",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                },
                "x-sse": true
            }
        },
        "/mobile/v1/message": {
            "get": {
                "security": [
                    {
                        "MobileToken": []
                    }
                ],
                "description": "Returns list of pending messages",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device",
                    "Messages"
                ],
                "summary": "Get messages for sending",
                "parameters": [
                    {
                        "enum": [
                            "lifo",
                            "fifo"
                        ],
                        "type": "string",
                        "default": "lifo",
                        "description": "Message processing order: lifo (default) or fifo",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of pending messages",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/smsgateway.MobileMessage"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "MobileToken": []
                    }
                ],
                "description": "Updates message state",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device",
                    "Messages"
                ],
                "summary": "Update message state",
                "parameters": [
                    {
                        "description": "List of message state updates",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/smsgateway.MobilePatchMessageItem"
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully updated"
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/mobile/v1/settings": {
            "get": {
                "security": [
                    {
                        "MobileToken": []
                    }
                ],
                "description": "Returns settings for a device",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device",
                    "Settings"
                ],
                "summary": "Get settings",
                "responses": {
                    "200": {
                        "description": "Settings",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.DeviceSettings"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/mobile/v1/user/code": {
            "get": {
                "security": [
                    {
                        "ApiAuth": []
                    }
                ],
                "description": "Returns one-time code for device registration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "Get one-time code for device registration",
                "responses": {
                    "200": {
                        "description": "User code",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.MobileUserCodeResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/mobile/v1/user/password": {
            "patch": {
                "security": [
                    {
                        "MobileToken": []
                    }
                ],
                "description": "Changes the user's password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "Change password",
                "parameters": [
                    {
                        "description": "Password change request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/smsgateway.MobileChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Password changed successfully"
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/mobile/v1/webhooks": {
            "get": {
                "security": [
                    {
                        "MobileToken": []
                    }
                ],
                "description": "Returns list of registered webhooks for device",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device",
                    "Webhooks"
                ],
                "summary": "List webhooks",
                "responses": {
                    "200": {
                        "description": "Webhook list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/smsgateway.Webhook"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "smsgateway.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code",
                    "type": "integer"
                },
                "data": {
                    "description": "Error context"
                },
                "message": {
                    "description": "Error message",
                    "type": "string",
                    "example": "An error occurred"
                }
            }
        },
...
        "smsgateway.MobileChangePasswordRequest": {
            "type": "object",
            "required": [
                "currentPassword",
                "newPassword"
            ],
            "properties": {
                "currentPassword": {
                    "description": "Current password",
                    "type": "string",
                    "example": "cp2pydvxd2zwpx"
                },
                "newPassword": {
                    "description": "New password, at least 14 characters",
                    "type": "string",
                    "minLength": 14,
                    "example": "cp2pydvxd2zwpx"
                }
            }
        },
        "smsgateway.MobileDeviceResponse": {
            "type": "object",
            "properties": {
                "device": {
                    "description": "Device information, empty if device is not registered on the server",
                    "allOf": [
                        {
                            "$ref": "#/definitions/smsgateway.Device"
                        }
                    ]
                },
                "externalIp": {
                    "description": "External IP",
                    "type": "string"
                }
            }
        },
        "smsgateway.MobileMessage": {
            "type": "object",
            "required": [
                "phoneNumbers"
            ],
            "properties": {
                "createdAt": {
                    "description": "Message creation time",
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                },
                "dataMessage": {
                    "description": "Data message",
                    "allOf": [
                        {
                            "$ref": "#/definitions/smsgateway.DataMessage"
                        }
                    ]
                },
                "deviceId": {
                    "description": "Optional device ID for explicit selection",
                    "type": "string",
                    "maxLength": 21,
                    "example": "PyDmBQZZXYmyxMwED8Fzy"
                },
                "id": {
                    "description": "ID (if not set - will be generated)",
                    "type": "string",
                    "maxLength": 36,
                    "example": "PyDmBQZZXYmyxMwED8Fzy"
                },
                "isEncrypted": {
                    "description": "Is encrypted",
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "description": "Message content\nDeprecated: use TextMessage instead",
                    "type": "string",
                    "maxLength": 65535,
                    "example": "Hello World!"
                },
                "phoneNumbers": {
                    "description": "Recipients (phone numbers)",
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "79990001234"
                    ]
                },
                "priority": {
                    "description": "Priority, messages with values greater than `99` will bypass limits and delays",
                    "default": 0,
                    "maximum": 127,
                    "minimum": -128,
                    "allOf": [
                        {
                            "$ref": "#/definitions/smsgateway.MessagePriority"
                        }
                    ],
                    "example": 0
                },
                "simNumber": {
                    "description": "SIM card number (1-3), if not set - default SIM will be used",
                    "type": "integer",
                    "maximum": 3,
                    "example": 1
                },
                "textMessage": {
                    "description": "Text message",
                    "allOf": [
                        {
                            "$ref": "#/definitions/smsgateway.TextMessage"
                        }
                    ]
                },
                "ttl": {
                    "description": "Time to live in seconds (conflicts with `validUntil`)",
                    "type": "integer",
                    "minimum": 5,
                    "example": 86400
                },
                "validUntil": {
                    "description": "Valid until (conflicts with `ttl`)",
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                },
                "withDeliveryReport": {
                    "description": "With delivery report",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "smsgateway.MobilePatchMessageItem": {
            "type": "object",
            "required": [
                "id",
                "recipients",
                "state"
            ],
            "properties": {
                "id": {
                    "description": "Message ID",
                    "type": "string",
                    "maxLength": 36,
                    "example": "PyDmBQZZXYmyxMwED8Fzy"
                },
                "recipients": {
                    "description": "Recipients states",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/smsgateway.RecipientState"
                    }
                },
                "state": {
                    "description": "State",
                    "allOf": [
                        {
                            "$ref": "#/definitions/smsgateway.ProcessingState"
                        }
                    ],
                    "example": "Pending"
                },
                "states": {
                    "description": "History of states",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "smsgateway.MobileRegisterRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "Device name",
                    "type": "string",
                    "maxLength": 128,
                    "example": "Android Phone"
                },
                "pushToken": {
                    "description": "FCM token",
                    "type": "string",
                    "maxLength": 256,
                    "example": "gHz-T6NezDlOfllr7F-Be"
                }
            }
        },
        "smsgateway.MobileRegisterResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "New device ID",
                    "type": "string",
                    "example": "QslD_GefqiYV6RQXdkM6V"
                },
                "login": {
                    "description": "User login",
                    "type": "string",
                    "example": "VQ4GII"
                },
                "password": {
                    "description": "User password, empty for existing user",
                    "type": "string",
                    "example": "cp2pydvxd2zwpx"
                },
                "token": {
                    "description": "Device access token",
                    "type": "string",
                    "example": "bP0ZdK6rC6hCYZSjzmqhQ"
                }
            }
        },
        "smsgateway.MobileUpdateRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID",
                    "type": "string",
                    "example": "QslD_GefqiYV6RQXdkM6V"
                },
                "pushToken": {
                    "description": "FCM token",
                    "type": "string",
                    "maxLength": 256,
                    "example": "gHz-T6NezDlOfllr7F-Be"
                }
            }
        },
        "smsgateway.MobileUserCodeResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "One-time code",
                    "type": "string",
                    "example": "123456"
                },
                "validUntil": {
                    "description": "One-time code expiration time",
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiAuth": {
            "type": "basic"
        },
        "MobileToken": {
            "description": "Mobile device token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ServerKey": {
            "description": "Private server authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "UserCode": {
            "description": "User one-time code authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}